Clippy needs Python installed as well
There seems to be a bit of a race condition in the CI tests. It
seems that the clippy test fails if Python is not installed yet,
but will pass if another test installs it first. I am unsure which
tests need Python, so I'll just add it to clippy for now and I
can add it to others later if needed.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>